image upload file service completed

+++ api/src/main/java/com/kullad/services/impl/FileStorageServiceImpl.java
@@ -1,5 +1,6 @@
 package com.kullad.services.impl;
 
+import com.kullad.exceptions.ResourceAlreadyExistsException;
 import com.kullad.exceptions.ResourceNotFoundException;
 import com.kullad.services.FileStorageService;
 import lombok.extern.slf4j.Slf4j;
@@ -21,18 +22,22 @@ public class FileStorageServiceImpl implements FileStorageService {
     private String uploadDir;
 
     @Override
-    public String storeFile(MultipartFile file) {
+    public String storeImageFile(MultipartFile file) {
         try {
             Path uploadPath = Paths.get(uploadDir);
             if (!Files.exists(uploadPath)) {
                 Files.createDirectory(uploadPath);
             }
 
+            String fileName = file.getOriginalFilename();
             Path filePath = uploadPath.resolve(Objects.requireNonNull(file.getOriginalFilename()));
+            if (Files.exists(filePath)) {
+                throw new ResourceAlreadyExistsException("Image file already exists with name: " + fileName);
+            }
             Files.copy(file.getInputStream(), filePath);
 
-            log.info("File uploaded successfully: {}", filePath);
-            return filePath.toString();
+            log.info("Image file uploaded successfully: {}", filePath);
+            return fileName;
         } catch (Exception e) {
             log.error("Error uploading image to server: {}", e.getMessage());
             throw new RuntimeException("Error uploading image to server: " + e.getMessage());
@@ -40,15 +45,15 @@ public class FileStorageServiceImpl implements FileStorageService {
     }
 
     @Override
-    public String getFile(String fileName) {
+    public byte[] getImageFile(String fileName) {
         try {
             Path filePath = Paths.get(uploadDir).resolve(fileName);
 
             if (Files.exists(filePath)) {
-                log.info("File retrieved successfully: {}", filePath);
-                return filePath.toString();
+                log.info("Image file retrieved successfully: {}", filePath);
+                return Files.readAllBytes(filePath);
             } else {
-                throw new ResourceNotFoundException("File not found: " + fileName);
+                throw new ResourceNotFoundException("Image file not found: " + fileName);
             }
         } catch (Exception e) {
             log.error("Error retrieving file: {}", e.getMessage());
@@ -57,7 +62,7 @@ public class FileStorageServiceImpl implements FileStorageService {
     }
 
     @Override
-    public String deleteFile(String fileName) {
+    public String deleteImageFile(String fileName) {
         try {
             Path filePath = Paths.get(uploadDir).resolve(fileName);
 
@@ -66,7 +71,7 @@ public class FileStorageServiceImpl implements FileStorageService {
                 log.info("File deleted successfully: {}", filePath);
                 return "File deleted successfully: " + fileName;
             } else {
-                throw new ResourceNotFoundException("File not found: " + fileName);
+                throw new ResourceNotFoundException("Image file not found: " + fileName);
             }
         } catch (IOException e) {
             log.error("Error deleting file: {}", e.getMessage());
